+++ hallo_eltern_cli/mda.py
@@ -1,7 +1,0 @@
-# This file is part of hallo-eltern-cli and licensed under the
-# Apache License Version 2.0 (See LICENSE.txt)
-# SPDX-License-Identifier: Apache-2.0
-
-class MDA(object):
-    def deliver(self, message):
-        pass